+++ polars_hist_db/loaders/__init__.py
@@ -1,6 +1,6 @@
-from .file.dsv_loader import load_typed_dsv
-from .file.file_search import find_files
-from .file.ziptools import convert_zipped_csvs_to_parquet
+from .dsv.dsv_loader import load_typed_dsv
+from .dsv.file_search import find_files
+from .dsv.ziptools import convert_zipped_csvs_to_parquet
 
 __all__ = [
     "load_typed_dsv",
